Add SweetAlert confirmation for lesson delete

+++ src/pages/MyLessons/MyLessonsTable/MyLessonsTable.jsx
@@ -5,6 +5,7 @@ import LoadingSpinner from "../../../components/LoadingSpinner";
 import useLessonHook from "../../../hooks/useLessonHook";
 import { toast } from "react-toastify";
 import { NavLink } from "react-router";
+import Swal from "sweetalert2";
 
 const MyLessonsTable = () => {
   const { user: firebaseUser } = useUserHook();
@@ -30,13 +31,25 @@ const MyLessonsTable = () => {
 
   //..................Delete a lesson from db by id.......................//
   const handleDelete = async (id) => {
-    const deleted = await deleteLesson(id);
+    const result = await Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#10B981", 
+      cancelButtonColor: "#EF4444", 
+      confirmButtonText: "Yes, delete it!",
+      cancelButtonText: "Cancel",
+    });
 
-    if (deleted) {
-      toast.success("Lesson deleted successfully");
-      refetch();
-    } else {
-      toast.error("Failed to delete lesson");
+    if (result.isConfirmed) {
+      const deleted = await deleteLesson(id);
+      if (deleted) {
+        toast.success("Lesson deleted successfully");
+        refetch();
+      } else {
+        toast.error("Failed to delete lesson");
+      }
     }
   };
